refactor : GridCell , ListCell 수정
- 이미지 처리관련 코드 추가
- 불필요한 로직 제거
